Maze has floors now and is winnable
Maze now has 5 floors!  I left all the doors open in case you guys
wanted to go through it.  The place to make all doors closed is in the
Wall class (I left a comment on the line you should change).

All door types are now implemented as well, and if you click the
treasure "door" you will get a message box saying you won.  We will need
to add more end of game stuff there at some point.

I still need to lock things down when a door gets clicked, but the maze
is now ready for Eric to do the minimap for it.

Note to Eric: for drawing the exit on the minimap, I threw in a static
point I made because it called on things that are not available any
more.  The logic would need to be reworked based on the changes I had to
make to make the maze have floors.  We can talk about this tomorrow.