Added front end difficulty stuff
Added the necessary parts to the main window and game core to allow for
a difficulty level to be chosen.

I also made it so the user must click to start a new game before a game
is playable.  Eric will be adding some safety stuff to disable movement
before a game is started later today.