locked movement when question is being asked
It is no longer possible to move around the dungeon if you have been
asked a question.  You are given an explanation as to why you don't move
if you click the door to try to move while being asked a question, but
my logic is wrong for the keyboard movement.  Ill take a look at it
again tomorrow when I am less out of it.