Fixed, Questions are now random
-Problem, it took me a long time to figure this out. Remember that
random takes off the internal clock so if you call it very fast it will
be the same number so when testing one by one to try and figure how why
the questions are all the same it seemed normal. but when you run the
game all the questions for eveyr door are the same. To fix this I amde a
RandomProvider class that threads off random isntances and makes a true
random number.

FINALY