open/closed doors show
So I got the floor to show which doors have been opened, closed, or
haven't been tried yet.

Nathan in the method the GetVisited did what I believe you were
expecting it do by not printing a red square where there is a wall for
the stairs but on a game load it causes the previously opened/locked
doors not to show as well.

Now that I have this method working, my super method can be removed
because after each question the method to find the open/closed doors
will execute and fill the door in with the appropriate color. All code
still remains in a blocked form, and I will delete it if we agree we
don't need it.