Mini-map fully functional
I've made the mini-map and floor buttons work as I had hoped they would.
I haven't removed any code, so if there is a serious problem with the
way I've set it up, we can talk about alternatives.

Some of the code is a little messy and I broke my own rules about using
the GetMaze method, but this is the best I could do in so short a time.

Since harder difficulties of the maze have some walls that don't have
doors, I commented out the lines that always painted every single door
regardless of whether there was one there or not.

It should be stressed that I used a lot of Eric's logic in getting this
to work.  I copy pasted a lot of the stuff from his other methods, just
used it slightly differently.