Question Database Tool user friendliness
Added a close button to the tool, a help button to explain everything to
the user, and made the window easily navigable via the keyboard.

I've not personally tested the Sound/Picture questions, but Daniel says
they work, so I believe this wraps up the coding side of the project.